+++ src/components/notebooklm/index.ts
@@ -1,0 +1,4 @@
+export { NotebookLMSidebar } from "./NotebookLMSidebar";
+export { NotebookLMChat } from "./NotebookLMChat";
+export { NotebookLMStudio } from "./NotebookLMStudio";
+export * from "./artifacts";